src/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.3
pydantic==2.10.2
requests==2.32.3
python-dateutil==2.9.0.post0
redis==5.1.1
